Split UserForm component in small components
