+++ frontend/app/users/page.tsx
@@ -1,7 +1,7 @@
 import { usersCollection } from "@/lib/mongodb";
 import authenticate from "../auth";
-import UserList from "@/components/UserList";
-import { User } from "./types";
+import UserList from "@/components/user/UserList";
+import { User } from "@/components/user/types";
 
 const page = async () => {
   const users = (await usersCollection.find().toArray()) as User[];
