+++ frontend/components/Inputs.tsx
@@ -1,0 +1,108 @@
+import React, { ChangeEvent } from "react";
+
+interface TextInputProps {
+  label: string;
+  value: string;
+  onChange: (e: ChangeEvent<HTMLInputElement>) => void;
+  required?: boolean;
+}
+
+export const TextInput: React.FC<TextInputProps> = ({
+  label,
+  value,
+  onChange,
+  required,
+}) => (
+  <div>
+    <label className="block text-sm font-medium text-gray-700">{label}</label>
+    <input
+      type="text"
+      value={value}
+      onChange={onChange}
+      className="border border-gray-300 p-3 rounded w-full focus:outline-none focus:ring-2 focus:ring-blue-500"
+      required={required}
+    />
+  </div>
+);
+
+interface NumberInputProps {
+  label: string;
+  value: number | string;
+  required?: boolean;
+  onChange: (e: ChangeEvent<HTMLInputElement>) => void;
+}
+
+export const NumberInput: React.FC<NumberInputProps> = ({
+  label,
+  value,
+  onChange,
+  required,
+}) => (
+  <div>
+    <label className="block text-sm font-medium text-gray-700">{label}</label>
+    <input
+      type="number"
+      value={value}
+      onChange={onChange}
+      className="border border-gray-300 p-3 rounded w-full focus:outline-none focus:ring-2 focus:ring-blue-500"
+      required={required}
+    />
+  </div>
+);
+
+interface SelectInputProps {
+  label: string;
+  value: string;
+  options: { value: string; label: string }[];
+  onChange: (e: ChangeEvent<HTMLSelectElement>) => void;
+  required?: boolean;
+}
+
+export const SelectInput: React.FC<SelectInputProps> = ({
+  label,
+  value,
+  options,
+  onChange,
+  required,
+}) => (
+  <div>
+    <label className="block text-sm font-medium text-gray-700">{label}</label>
+    <select
+      value={value}
+      onChange={onChange}
+      className="border border-gray-300 p-3 rounded w-full focus:outline-none focus:ring-2 focus:ring-blue-500"
+      required={required}
+    >
+      {options.map((option) => (
+        <option key={option.value} value={option.value}>
+          {option.label}
+        </option>
+      ))}
+    </select>
+  </div>
+);
+
+interface CheckboxInputProps {
+  label: string;
+  checked: boolean;
+  onChange: (e: ChangeEvent<HTMLInputElement>) => void;
+  required?: boolean;
+}
+
+export const CheckboxInput: React.FC<CheckboxInputProps> = ({
+  label,
+  checked,
+  onChange,
+  required,
+}) => (
+  <div>
+    <label className="block text-sm font-medium text-gray-700">{label}</label>
+    <input
+      type="checkbox"
+      checked={checked}
+      onChange={onChange}
+      className="p-3 focus:outline-none focus:ring-2 focus:ring-blue-500"
+      required={required}
+    />
+  </div>
+);
