+++ frontend/components/constants.ts
@@ -1,0 +1,21 @@
+export const TIME_FRAME_OPTIONS = [
+  { value: "minute", label: "minute" },
+  { value: "day", label: "day" },
+  { value: "3minute", label: "3minute" },
+  { value: "5minute", label: "5minute" },
+  { value: "10minute", label: "10minute" },
+  { value: "15minute", label: "15minute" },
+  { value: "30minute", label: "30minute" },
+  { value: "60minute", label: "60minute" },
+];
+
+export const TRADE_TYPE_OPTIONS = [
+  { value: "BUY", label: "BUY" },
+  { value: "SELL", label: "SELL" },
+  { value: "BOTH", label: "BOTH" },
+];
+
+export const BROKER_OPTIONS = [
+  { value: "AngelOne", label: "AngelOne" },
+  { value: "Zerodha", label: "Zerodha" },
+];
